% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgGDP.R
\name{calcAgGDP}
\alias{calcAgGDP}
\title{calcAgGDP}
\usage{
calcAgGDP()
}
\value{
List of magpie object with results on country  level, weight, unit and description.
}
\description{
calculates Ag GDP as \% of total GDP
}
\examples{

\dontrun{
calcOutput("AgGDP")
}
}
\author{
David Chen
}
