% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPopulation.R
\name{calcValidPopulation}
\alias{calcValidPopulation}
\title{calcValidPopulation}
\usage{
calcValidPopulation(datasource = "WDI")
}
\arguments{
\item{datasource}{Various daasource to load population data from can be specified here. So far available: WDI (default)}
}
\description{
Returns historical development of population
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich, Kristine Karsens
}
