% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTutorialWDI.R
\name{convertTutorialWDI}
\alias{convertTutorialWDI}
\title{Convert WDI}
\usage{
convertTutorialWDI(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing WDI data region resolution}

\item{subtype}{Name of the worldbank indicator, e.g. "SP.POP.TOTL"}
}
\value{
MAgPIE object of the WDI data disaggregated to country level
}
\description{
Convert WDI converts data from readWDI() to ISO country level. Adds Taiwan
as difference from global total.
}
\examples{

\dontrun{ a <- convertWDI(x)
}
}
\author{
Jan Phillip Dietrich, Benjamin Bodirsky, Xiaoxi Wang, David Chen
}
