% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTutorialWDI.R
\name{downloadTutorialWDI}
\alias{downloadTutorialWDI}
\title{Dowload WDI}
\usage{
downloadTutorialWDI()
}
\description{
Download WDI (World development indicators) data .rda file.
}
\examples{

\dontrun{ a <- downloadSource(type="TutorialWDI")
}

}
\seealso{
\code{\link[madrat]{downloadSource}} \code{\link[WDI]{WDI}}
}
\author{
David Chen
}
