% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTUTORIAL.R
\name{fullTUTORIAL}
\alias{fullTUTORIAL}
\title{fullMAgPIE}
\usage{
fullTUTORIAL(rev = numeric_version("1"), dev = "", extra = "Example Argument")
}
\arguments{
\item{rev}{data revision which should be used as input (numeric_version).}

\item{dev}{dev flag for testing}

\item{extra}{extra text to insert}
}
\description{
Function that produces the complete regional data set required for running the
MAgPIE model.
}
\examples{
\dontrun{
retrieveData("tutorial", rev = numeric_version("12"),
             mainfolder = "pathtowhereallfilesarestored")
}
}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}}, \code{\link[madrat]{calcOutput}}
}
\author{
David M Chen
}
