% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTUTORIALVALIDATION.R
\name{fullTUTORIALVALIDATION}
\alias{fullTUTORIALVALIDATION}
\title{fullTUTORIALVALIDATION}
\usage{
fullTUTORIALVALIDATION(rev = 0.1)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}
}
\description{
Function that produces the complete validation data set used for evaluation of MAgPIE outputs
}
\examples{
\dontrun{
retrieveData("VALIDATION")
}
}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}}, \code{\link[madrat]{calcOutput}}
}
\author{
Jan Philipp Dietrich, Kristine Karstens
}
