% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtutorial-package.R
\docType{package}
\name{mrtutorial-package}
\alias{mrtutorial}
\alias{mrtutorial-package}
\title{mrtutorial: Tutorial Package For Madrat Package Library}
\description{
Example package of mr- world libraries, in conjunction with MAgPIE MADRat tutorial. For more information please see: https://github.com/magpiemodel/tutorials/tree/master/madrat .
}
\author{
\strong{Maintainer}: David Meng-Chuen Chen \email{david.chen@pik-potsdam.de}

}
