% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTutorialWDI.R
\name{downloadTutorialWDI}
\alias{downloadTutorialWDI}
\title{Dowload WDI}
\usage{
downloadTutorialWDI()
}
\description{
Download WDI (World development indicators) data .rda file.
the WDI data is updated with the funciton "WDISearch(cache=WDIcache())"
}
\examples{

\dontrun{ a <- downloadSource(type="WDI")
}

}
\seealso{
\code{\link{downloadSource}} \code{\link{WDI}}
}
\author{
David Chen
}
